% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_bathing_quality_eu.R
\name{assess_bathing_quality_eu}
\alias{assess_bathing_quality_eu}
\title{Bathing quality assessment (EU)}
\usage{
assess_bathing_quality_eu(e.coli, log = TRUE)
}
\arguments{
\item{e.coli}{A numeric vector with e.coli values}

\item{log}{logical. Are the values log-values?}
}
\value{
Returns a single factor with all quality levels
}
\description{
Computes the quality assessment according to european bathing directive
2006/7/EC from E.Coli values. The four possible quality levels are:
excellent, good, sufficient and poor
}
