% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\alias{ask_for_variables}
\alias{process_model_riverdata}
\alias{create_formula}
\title{Build a Model for E.Coli}
\usage{
build_model(riverdata, variables = ask_for_variables(riverdata))

ask_for_variables(riverdata)

process_model_riverdata(riverdata, variables)

create_formula(variables)
}
\arguments{
\item{riverdata}{a list with riverdata (hygiene + physical data)}

\item{variables}{character. Selected variables for the model}
}
\value{
Returns a model of the riverdata.

Returns a character-vector with the chosen model variables

Returns a data.frame with data for hygiene and chosen variables

Returns parsed model-formula. (Like model$formula)
}
\description{
Functions for modelbuilding \\n
\code{build_model}: takes the riverdata, handles the other functions and
invokes \code{\link[rstanarm]{stan_lm}}
}
\details{
Build the model from hygiene data and physical data like flow, rain,
 wwtp. Asks for user input to select variables. Computes the data.frame with
 data for hygiene and chosen variables and creates a formula of the form:
\code{Q*(K + R)} while multiple Qs will be multiplied, multiple Ks and Rs
will be added.
}
\section{Functions}{
\itemize{
\item \code{ask_for_variables}: Internal function. Quite time consuming

\item \code{process_model_riverdata}: Internal usage

\item \code{create_formula}: Internal usage
}}

\examples{
\dontrun{variables <- c("e.coli","q_havel",...)
lm(formula = eval(create_formula(variables)),
data = process_model_riverdata(riverdata, variables))}

create_formula(c("log_e.coli","q_havel","ka_ruhleben","r_berlin"))
create_formula(c("e.coli","r_mitte","r_charlottenburg","r_spandau"))

}
