% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_new_model.R
\name{build_new_model}
\alias{build_new_model}
\title{Create a new Model}
\usage{
build_new_model(river)
}
\arguments{
\item{river}{character. river you want to build a model on}
}
\value{
This function returns merely a message what happend.
}
\description{
Main function for creation, object handling and saving of new models.
The models will be saved in a own subdirectory as r-objects as a side-effect.
The models will not be returned. They have to be loaded by other functions.
}
