% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_model.R
\name{choose_model}
\alias{choose_model}
\title{Set User Input for a List of Models}
\usage{
choose_model(rivermodels)
}
\arguments{
\item{rivermodels}{A list of named models}
}
\value{
Returns the user input as character vector, or an ERROR message.
}
\description{
Choose model asks the user inside the console for input. Options are Exit,
New Model, or one of a list of existing models. If no integer number was
presented by the user an ERROR message will be created but no ERROR will be
thrown. This way this can be inserted inside a loop.
}
\examples{
\donttest{choose_model(list())}
\donttest{choose_model(list(fake_river_model = 1))}

}
