% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_variables.R
\name{correlation_scatterplot}
\alias{correlation_scatterplot}
\alias{correlation_values}
\title{Scatterplotmatrix of similar Variables to E.Coli}
\usage{
correlation_scatterplot(df, ...)

correlation_values(df, ...)
}
\arguments{
\item{df}{data.frame with data for e.coli and chosen variables in lagdays}

\item{\dots}{Arguments passed to \code{stats::cor}}
}
\value{
Plotting function. Returns a plot.

Returns correlation values.
}
\description{
Takes similar named variables and produces a matrix of scatterplots and their
correlation coefficients to E.Coli.
}
\section{Functions}{
\itemize{
\item \code{correlation_values}: Internal function
}}

\examples{
\donttest{correlation_values(data.frame(datum = rep("egal",10), e.coli = 1:10, var = 1:10), variable = "var")}

}
