% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_paths}
\alias{get_paths}
\title{Get List of Paths used in the Flusshygiene Project}
\usage{
get_paths(resolve = TRUE, ...)
}
\arguments{
\item{resolve}{if \code{TRUE} (default) path placeholders are resolved}

\item{\dots}{arguments passed to \code{\link[kwb.utils]{resolve}} if
\code{resolve} is \code{TRUE}}
}
\description{
Get List of Paths used in the Flusshygiene Project
}
\examples{
\dontrun{
paths <- get_paths()

# Paths to the different work package folders
paths$ap2
paths$ap3
paths$ap4

# What tables are contained in the ODM database?
kwb.db::hsTables(paths$odm)

# Get all Flusshygiene data into one data frame
data <- kwb.ogre.model::get_lab_values(paths$odm)
}

}
