% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_riverdata.R
\name{import_riverdata}
\alias{import_riverdata}
\title{Read existing River Data}
\usage{
import_riverdata(path)
}
\arguments{
\item{path}{character-string to a DATA_preprocessed_csv directory}
}
\value{
Returns a list of data.frames containing the river data
}
\description{
Read existing, preprocessed csv files with first column datetime and other 
columns variable information.
}
