% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwb.flusshygiene.R
\docType{package}
\name{kwb.flusshygiene}
\alias{kwb.flusshygiene}
\alias{kwb.flusshygiene-package}
\title{A package for creating and handling river hygiene models}
\description{
The kwb.flusshygiene package provides functions in three major categories:
model handling, model creation and model prediction.
}
\section{Model handling}{
 \describe{
  \item{\code{\link{river_model_prediction}}}{is the main function in this
  package. It uses all of the following functions from within.}
  \item{\code{\link{get_paths}}}{reads a serverpath library for easy
  directory accessing} \item{\code{\link{search_existing_models}}}{searches
  saved R-objects in the river directories.} }
}

\section{Model creation}{
 \describe{
  \item{\code{\link{build_new_model}}}{again a overhead function for model
  creation. Asks the user whether or not the new model shall be saved as
  R-object.} \item{\code{\link{import_riverdata}}}{ reads all river data from
  a directory.} \item{\code{\link{build_model}}}{ is a small function
  handling model creation and invokes \code{\link[rstanarm]{stan_lm}} for
  model building.} \item{\code{\link{ask_for_variables}}}{asks the user which
  variables shall be included in the model. Creates plots as a side effect.}
  \item{\code{\link{process_model_riverdata}}}{processes a data.frame with
  the necessary data for the data argument in
  \code{\link[rstanarm]{stan_lm}}}
  \item{\code{\link{create_formula}}}{creates a hygiene formula out of the
  variables with the form \code{e.coli ~ Q * (R + Ka)}} }
}

\section{Model prediction}{
 \describe{ \item{\code{\link{predict_quality}}}{is
  the overhead function for the prediction. It also invokes
  \code{\link[rstanarm]{posterior_predict}}}
  \item{\code{\link{get_newdata}}}{gathers the latest data for prediction.}
  \item{\code{\link{print_latest}}}{prints the prediction of the latest day.}
  \item{\code{\link{plot_predicted_quality}}}{plots a whole season with
  quality assessment.} }
}

\section{Utility functions}{
 \describe{
  \item{\code{\link{unroll_physical_data}}}{unrolls a list with data with
  lagday combinations to 5 days (default)}
  \item{\code{\link{correlation_scatterplot}}}{plots a scatterplot matrix of
  the unrolled physical data together with correlation values.} }
}

\section{Plotting functions}{
 \describe{
  \item{\code{\link{plot_stan_model}}}{plots a stan_lm posterior predction
  with quality assessment.} \item{\code{\link{plot_data_overview}}}{plot data
  overview} \item{\code{\link{plot_hygiene_overview}}}{a statistical hygiene
  data overview} \item{\code{\link{plot_q_overview}}}{plot a overview of all
  q values} \item{\code{\link{plot_rain_overview}}}{ plot a monthly overview
  of all rain gauges} }
}

