% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_overview.R
\name{plot_data_overview}
\alias{plot_data_overview}
\title{Plot Data Timeline Overview}
\usage{
plot_data_overview(riverdata, type = "segment")
}
\arguments{
\item{riverdata}{A list of hygiene and physical data of the river}

\item{type}{Either "segment" or "point" for more precise information}
}
\value{
Returns a plot
}
\description{
Creates a plot with segments or points of the data availability.
}
