% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hygiene_overview.R
\name{plot_hygiene_overview}
\alias{plot_hygiene_overview}
\title{Plot Hygiene Overview}
\usage{
plot_hygiene_overview(hygiene_df)
}
\arguments{
\item{hygiene_df}{A data.frame with the hygiene data of a given river}
}
\value{
Returns a plot
}
\description{
Creates a plot with three graphs: Histogramm of all e.coli values, a density
curve of the last 16 values, and a boxplot of all values again
}
