% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prediction.R
\name{plot_predicted_quality}
\alias{plot_predicted_quality}
\title{Plot Quality}
\usage{
plot_predicted_quality(model, prediction, ...)
}
\arguments{
\item{model}{stan.lm model for the river}

\item{prediction}{list of season, ppd of predcit and ppd of means}

\item{...}{Further parameter passed to plot.default}
}
\value{
Plotting function. Returns a plot.
}
\description{
Window function for \link{plot_stan_model}
}
