% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_q_overview.R
\name{plot_q_overview}
\alias{plot_q_overview}
\title{Plot Flowing Conditions}
\usage{
plot_q_overview(q_df)
}
\arguments{
\item{q_df}{The data.frame with 2 columns: datum and q}
}
\value{
Returns a plot
}
\description{
Creates a plot with the standard flowing conditions over the year. The data
of all years will be taken into account.
}
