% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rain_overview.R
\name{plot_rain_overview}
\alias{plot_rain_overview}
\title{Plot Monthly Rain Summary}
\usage{
plot_rain_overview(df)
}
\arguments{
\item{df}{A data frame with different rain gauges.}
}
\value{
Returns a plot
}
\description{
Creates a plot with a monthly summary overview over the different rain sites
}
