% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stan_model.R
\name{plot_stan_model}
\alias{plot_stan_model}
\title{Plot Model Prediction with Quality Assessment}
\usage{
plot_stan_model(timestamp, predict, linpred, log = FALSE, q90, q95,
  nlines = 250, nlinesCenter = 100, ...)
}
\arguments{
\item{timestamp}{POSIX. The x-axis timestamp}

\item{predict}{ppd. The posterior prediction of the model}

\item{linpred}{ppd. The linpred (predicted means) of the model}

\item{log}{logical. Is E.Coli log01-transformed?}

\item{q90}{numeric. The 90. percentile of predict.}

\item{q95}{numeric. The 95. precentile of predict.}

\item{nlines}{numeric. How many lines for posterion predictions?}

\item{nlinesCenter}{numeric. How many lines for predicted means?}

\item{...}{Further parameters for plot.default}
}
\value{
Plotting function. Returns a plot.
}
\description{
Plots a sample of posterior predictions and means. Furthermore colours an 
hygiene quality assessment as background (see EU Bathing Water Directive)
Dark blue means excellent quality.
Steelblue means good quality.
Yellow means sufficient quality.
Red means insufficient quality.
}
