% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prediction.R
\name{predict_quality}
\alias{predict_quality}
\alias{get_newdata}
\alias{print_latest}
\alias{get_latest_season}
\title{Predict Hygiene Quality}
\usage{
predict_quality(model, river_dir, output = "season")

get_newdata(variables, river_dir)

print_latest(model, newdata)

get_latest_season(newdata)
}
\arguments{
\item{model}{stan_lm. A model of e.coli concentration in given river}

\item{river_dir}{character. Path to river-data for up-to-date predictions.}

\item{output}{character. \code{"season"} will return a list with prediction,
\code{"latest"} will return console output}

\item{variables}{character. A vector with all variables used in the model}

\item{newdata}{data.frame with physical data used in the model}
}
\value{
Returns a list of physical data and prediction and linpred from model

Returns a data.frame with the merged data found
}
\description{
Main function for invoking and object handling. E.Coli hygiene models will
be used to predict hygiene quality on differnt scopes.
}
\section{Functions}{
\itemize{
\item \code{get_newdata}: Internal Usage

\item \code{print_latest}: Internal Usage

\item \code{get_latest_season}: Internal Usage
}}

