% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flusshygieneMain.R
\name{readTableData}
\alias{readTableData}
\title{Read Data for ODM Tables}
\usage{
readTableData(sourcedir = system.file("extdata", "ODM", package =
  "kwb.flusshygiene"))
}
\arguments{
\item{sourcedir}{path to input directory}
}
\value{
list of data frames
}
\description{
Read data for ODM tables from CSV files stored in the package
}
