% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_model_prediction.R
\name{river_model_prediction}
\alias{river_model_prediction}
\alias{search_existing_models}
\title{Programm for Model Handling and Prediction}
\usage{
river_model_prediction(river)

search_existing_models(river_dir)
}
\arguments{
\item{river}{character. The desired river, like "isar".}

\item{river_dir}{character. Path to server and river directory}
}
\value{
(invisible) The data.frame returned by the prediction plus a date
column for easy plotting.

Returns a list with the existing models for that river
(empty if no model was found).
}
\description{
This function is the front-end for model search on the server, model building
with existing data, or prediction with new data. It invokes all other
functions and handles their objects. It is a main function.
}
\section{Functions}{
\itemize{
\item \code{search_existing_models}: directory searching
}}

\examples{
\donttest{
river_model_prediction(river = "isar")
}

\donttest{
serverpath <- "//poseidon/projekte$/SUW_Department/Projects/FLUSSHYGIENE/Data-Work packages/Daten"
river_dir <- search_existing_river_dir(river = "isar", server = serverpath)
search_existing_models(river_dir = river_dir)}

}
