% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unroll_lagdays.R
\name{unroll_physical_data}
\alias{unroll_physical_data}
\alias{unroll_lagdays}
\title{Unroll Lagdays of Data}
\usage{
unroll_physical_data(physical_data)

unroll_lagdays(df, n = 5)
}
\arguments{
\item{physical_data}{list of river data (without hygiene)}

\item{df}{data.frame of 2 columns: datum and var}

\item{n}{numeric. unto to which day shall be lagged behind?}
}
\value{
Returns a list of data.frames for each variable. The data.frames
contain the unrolled lagdays (with maxday = 5, length(df) == 17)
}
\description{
Unrolls the lagdays of data.frames.
}
\section{Functions}{
\itemize{
\item \code{unroll_lagdays}: Internal usage mostly
}}

\examples{
df1 <- data.frame(datum = rep("egal", 25), var = 1:25)
df2 <- data.frame(datum = rep("egal", 25), var2 = 51:75, var3 = 101:125)
unroll_lagdays(df1)
summary(unroll_physical_data(list(var1 = df1, var2 = df2)))

}
